+++ backend/interfaces/api/handler/handler.go
@@ -1,0 +1,7 @@
+package handler
+
+import "github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
+
+type Handler struct {
+	*request.Context
+}
