+++ backend/interfaces/api/handler/lastest_articles.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/yoshihiro-shu/draft-backend/backend/application/usecase"
 	"github.com/yoshihiro-shu/draft-backend/backend/domain/model"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/pager"
 	"go.uber.org/zap"
