+++ backend/interfaces/api/handler/user.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/yoshihiro-shu/draft-backend/backend/application/usecase"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/auth"
 	"go.uber.org/zap"
 )
