--- backend/interfaces/api/server/handler/handler.go
@@ -1,7 +1,0 @@
-package handler
-
-import "github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
-
-type Handler struct {
-	*request.Context
-}
