+++ backend/interfaces/api/server/server.go
@@ -8,8 +8,8 @@ import (
 	"os/signal"
 	"time"
 
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/router"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/user_api"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/router"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/user_api"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/model"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
