+++ backend/interfaces/api/user_api/api.go
@@ -1,10 +1,10 @@
 package user_api
 
 import (
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/handler"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/middlewares"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/router"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/handler"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/middlewares"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/router"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/model"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/config"
