+++ backend/registory/top_page.go
@@ -3,8 +3,8 @@ package registory
 import (
 	"github.com/yoshihiro-shu/draft-backend/backend/application/usecase"
 	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/persistence"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/handler"
-	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/handler"
+	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/request"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
 	"gorm.io/gorm"
 )
